+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Upcoming release
 * Fix `sendPassword` & `presentScannable` invocations
+* Fix user preference parsing
 
 ## 2.2.7
 * Updated to Android SDK 2.2.3